Handle more kinds of proxy connection errors
Updates the exception handler for `ProxyConnectionHandler` to handle
more kinds of errors, not just `SocketError`.  The harm in retrying more
kinds of errors seems quite small, and the harm in *not* retrying is
rather large.

We've seen in production that `Errno::EPIPE`, which is not a
`SocketError`, is quite common.  Not attempting to reconnect the
connection in that case is quite bad, since that's not the kind of error
that resolves itself.